use DI for the event-recovery code as well
